Refactored Interop and started with host support. VstHostCommandProxy will take on calls from the plugin (audioMasterCallback), VstPluginCommandStub will dispatch calls to the plugin (AEffect funct. ptr's) and VstPluginContext contains all information on a specific plugin instance. The Initialize method will load the plugin dll and call the exported main function.

+++ Source/Code/Jacobi.Vst.Interop/Plugin/HostCommandStub.h
@@ -16,7 +16,7 @@ public:
 	virtual System::Boolean ProcessEvents(array<Jacobi::Vst::Core::VstEvent^>^ events);
 	virtual System::Boolean IoChanged();
 	virtual System::Boolean SizeWindow(System::Int32 width, System::Int32 height);
-	virtual System::Double GetSampleRate();
+	virtual System::Single GetSampleRate();
 	virtual System::Int32 GetBlockSize();
 	virtual System::Int32 GetInputLatency();
 	virtual System::Int32 GetOutputLatency();
